+++ packages/gatsby-theme-apollo-docs/gatsby-node.js
@@ -1,7 +1,7 @@
 const jsYaml = require('js-yaml');
 const path = require('path');
 const {createFilePath} = require('gatsby-source-filesystem');
-const {getVersionBasePath} = require('./src/utils');
+const {getVersionBasePath, generateNavItems} = require('./src/utils');
 
 async function onCreateNode({node, actions, getNode, loadNodeContent}) {
   if (configPaths.includes(node.relativePath)) {
@@ -158,7 +158,11 @@ exports.createPages = async ({actions, graphql}, options) => {
     spectrumPath,
     typescriptApiBox,
     versions = {},
-    defaultVersion
+    defaultVersion,
+    algoliaApiKey,
+    algoliaIndexName,
+    navConfig,
+    baseUrl
   } = options;
 
   const {edges} = data.allFile;
@@ -223,7 +227,11 @@ exports.createPages = async ({actions, graphql}, options) => {
         spectrumPath: spectrumPath || repo,
         typescriptApiBox,
         versions: versionKeys, // only need to send version labels to client
-        defaultVersion
+        defaultVersion,
+        algoliaApiKey,
+        algoliaIndexName,
+        navItems: generateNavItems(navConfig),
+        baseUrl
       }
     });
   });
