+++ packages/gatsby-theme-apollo-docs/index.js
@@ -9,3 +9,4 @@ export {
   ExpansionPanelList,
   ExpansionPanelListItem
 } from './src/components/expansion-panel';
+export {defaultOptions} from './src/default-options';
