+++ packages/gatsby-theme-apollo-docs/src/components/nav/index.js
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import styled from '@emotion/styled';
 import {breakpoints} from 'gatsby-theme-apollo';
+import {formatValue} from '../../utils';
 
 const Container = styled.nav({
   display: 'flex',
@@ -14,79 +15,10 @@ const Container = styled.nav({
   }
 });
 
-const navConfig = {
-  '/docs': {
-    text: 'Platform',
-    matchRegex: /^\/docs\/(intro|platform|resources|references|$)/
-  },
-  '/docs/tutorial/introduction': {
-    text: 'Tutorial',
-    matchRegex: /^\/docs\/tutorial/
-  },
-  '/docs/react': {
-    text: 'Client',
-    subpages: {
-      '/docs/react': 'React + React Native',
-      '/docs/angular': 'Angular',
-      'https://github.com/akryum/vue-apollo': 'Vue.js',
-      '/docs/link': 'Apollo Link',
-      '/docs/ios': 'Native iOS',
-      '/docs/android': 'Native Android',
-      '/docs/scalajs': 'Scala.js'
-    }
-  },
-  '/docs/apollo-server': {
-    text: 'Server',
-    subpages: {
-      '/docs/apollo-server': 'Apollo Server',
-      '/docs/graphql-tools': 'graphql-tools',
-      '/docs/graphql-subscriptions': 'GraphQL subscriptions'
-    }
-  },
-  '/docs/community': {
-    text: 'Community',
-    subpages: {
-      'https://blog.apollographql.com': 'Blog',
-      'https://spectrum.chat/apollo': 'Spectrum',
-      'https://twitter.com/apollographql': 'Twitter',
-      'https://youtube.com/channel/UC0pEW_GOrMJ23l8QcrGdKSw': 'YouTube',
-      '/docs/community': 'Contribute',
-      'https://summit.graphql.com': 'GraphQL Summit',
-      'https://graphql.com': 'Explore GraphQL'
-    }
-  }
-};
-
-function formatValue(value) {
-  return value.startsWith('/')
-    ? `https://www.apollographql.com${value}`
-    : value;
-}
-
-function generateSubpage([value, text]) {
-  return {
-    value,
-    text
-  };
-}
-
-function generateNavItems(config) {
-  return Object.entries(config).map(
-    ([value, {text, matchRegex, subpages}]) => ({
-      text,
-      value: formatValue(value),
-      matchRegex,
-      subpages: subpages && Object.entries(subpages).map(generateSubpage)
-    })
-  );
-}
-
-export const navItems = generateNavItems(navConfig);
-
 export default function Nav(props) {
   return (
     <Container>
-      {navItems.map(({value, text, matchRegex, subpages}) => {
+      {props.items.map(({value, text, matchRegex, subpages}) => {
         let isActive = matchRegex
           ? matchRegex.test(props.pathname)
           : props.isPathActive(value);
@@ -99,7 +31,7 @@ export default function Nav(props) {
         return (
           <NavItem
             key={value}
-            href={formatValue(value)}
+            href={formatValue(props.baseUrl, value)}
             subpages={subpages}
             active={isActive}
           >
@@ -112,6 +44,8 @@ export default function Nav(props) {
 }
 
 Nav.propTypes = {
+  items: PropTypes.array.isRequired,
+  baseUrl: PropTypes.string.isRequired,
   pathname: PropTypes.string.isRequired,
   isPathActive: PropTypes.func.isRequired
 };
